requirements.txt
djangorestframework
Django
Pygments
Markdown
coreapi
psycopg2-binary
dj-database-url
gunicorn
whitenoise
PyYAML
django-mysql
